+++ src/main/java/com/dasbiersec/reloader/repos/AccountRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends CrudRepository<Account, Integer>
+public interface AccountRepository extends CrudRepository<Account, Integer>
 {
 	Account findUserByUsername(String username);
 }