+++ src/main/java/com/dasbiersec/reloader/repos/ComponentRepository.java
@@ -10,4 +10,7 @@ import org.springframework.stereotype.Repository;
 public interface ComponentRepository extends CrudRepository<Component, Integer>
 {
 	Iterable<Component> findComponentByType(ComponentType type);
+	Iterable<Component> findAllByUserId(Integer userId);
+	Iterable<Component> findComponentByUserIdAndType(Integer userId, ComponentType type);
+	Component findOneByUserIdAndId(Integer userId, Integer id);
 }